# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PhotoEditor",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/library/current": {
      "put": {
        "tags": [
          "library"
        ],
        "summary": "Open Folder",
        "operationId": "open_folder_api_library_current_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenFolderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/import": {
      "post": {
        "tags": [
          "library"
        ],
        "summary": "Import Folder",
        "operationId": "import_folder_api_library_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fs/dirs": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Fs Dirs",
        "operationId": "fs_dirs_api_fs_dirs_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Folder to list; omit for the drives.",
              "title": "Path"
            },
            "description": "Folder to list; omit for the drives."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirListing"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/photos": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "List Photos",
        "operationId": "list_photos_api_photos_get",
        "parameters": [
          {
            "name": "style_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by style id; 'none' = photos without a style.",
              "title": "Style Id"
            },
            "description": "Filter by style id; 'none' = photos without a style."
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 0,
              "default": 0,
              "title": "Min Rating"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PhotoSort",
              "default": "date"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Photo_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/photos/{photo_id}": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Photo Detail",
        "operationId": "photo_detail_api_photos__photo_id__get",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/photos/{photo_id}/sidecar": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Photo Sidecar",
        "description": "The camera's own JPEG of a RAW (if it saved one), for comparing with the default look.",
        "operationId": "photo_sidecar_api_photos__photo_id__sidecar_get",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 4096,
              "minimum": 256,
              "description": "Long edge in pixels.",
              "default": 1600,
              "title": "Size"
            },
            "description": "Long edge in pixels."
          }
        ],
        "responses": {
          "200": {
            "description": "JPEG image",
            "content": {
              "image/jpeg": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/photos/{photo_id}/edit": {
      "put": {
        "tags": [
          "edit"
        ],
        "summary": "Save Edit",
        "description": "Save a photo's adjustments (the full set; only what differs from the defaults is stored).",
        "operationId": "save_edit_api_photos__photo_id__edit_put",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustmentParams-Input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "edit"
        ],
        "summary": "Reset Edit",
        "description": "Drop the photo's own tweaks (its style stays; PUT /style with null removes the style).",
        "operationId": "reset_edit_api_photos__photo_id__edit_delete",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/photos/{photo_id}/style": {
      "put": {
        "tags": [
          "edit"
        ],
        "summary": "Set Photo Style",
        "description": "Give the photo a style right away (null removes it). Tweaks of what the style sets are replaced.",
        "operationId": "set_photo_style_api_photos__photo_id__style_put",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoStyleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/photos/{photo_id}/thumbnail": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Photo Thumbnail",
        "operationId": "photo_thumbnail_api_photos__photo_id__thumbnail_get",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JPEG image",
            "content": {
              "image/jpeg": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/photos/{photo_id}/preview": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Photo Preview",
        "operationId": "photo_preview_api_photos__photo_id__preview_get",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show the unedited photo (same as after until Phase 3).",
              "default": false,
              "title": "Before"
            },
            "description": "Show the unedited photo (same as after until Phase 3)."
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 4096,
              "minimum": 256,
              "description": "Long edge in pixels.",
              "default": 1600,
              "title": "Size"
            },
            "description": "Long edge in pixels."
          }
        ],
        "responses": {
          "200": {
            "description": "JPEG image",
            "content": {
              "image/jpeg": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/styles": {
      "get": {
        "tags": [
          "styles"
        ],
        "summary": "List Styles",
        "operationId": "list_styles_api_styles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StyleSummary"
                  },
                  "type": "array",
                  "title": "Response List Styles Api Styles Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "styles"
        ],
        "summary": "Create Style",
        "operationId": "create_style_api_styles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/styles/from-photo": {
      "post": {
        "tags": [
          "styles"
        ],
        "summary": "Create Style From Photo",
        "description": "A new style from a photo's current look (chosen groups; exposure and white balance as rules).",
        "operationId": "create_style_from_photo_api_styles_from_photo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleFromPhoto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/styles/{style_id}/from-photo": {
      "post": {
        "tags": [
          "styles"
        ],
        "summary": "Update Style From Photo",
        "operationId": "update_style_from_photo_api_styles__style_id__from_photo_post",
        "parameters": [
          {
            "name": "style_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Style Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleUpdateFromPhoto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/export-presets/{preset_id}": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Preset",
        "operationId": "preset_api_export_presets__preset_id__get",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportPreset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LibraryInfo": {
        "properties": {
          "folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder",
            "description": "Currently opened photo folder."
          },
          "include_subfolders": {
            "type": "boolean",
            "title": "Include Subfolders",
            "description": "The Library also shows photos in subfolders.",
            "default": false
          },
          "photo_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Photo Count"
          },
          "suggested_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Folder",
            "description": "Folder to offer when nothing is open yet (the configured sample folder)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "folder",
          "include_subfolders",
          "photo_count",
          "suggested_folder"
        ],
        "title": "LibraryInfo"
      },
...
      "OpenFolderRequest": {
        "properties": {
          "folder": {
            "type": "string",
            "minLength": 1,
            "title": "Folder",
            "description": "An already imported folder to show in the Library."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "folder"
        ],
        "title": "OpenFolderRequest"
      },
...
      "PhotoSort": {
        "type": "string",
        "enum": [
          "date",
          "name",
          "rating"
        ],
        "title": "PhotoSort"
      },
...
    }
  }
}